camera_supervisor_requirements.txt
opencv-python
onvif-zeep
numpy
PyYAML
python-ffmpeg
ultralytics
boto3
paho-mqtt
sib-api-v3-sdk
